requirements.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
fastapi==0.79.0
fonttools==4.34.4
h11==0.13.0
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
orjson==3.7.12
